MEAS.MNT: Move energymeas from csdevice and refactor measurements subpackage.
