+++ siriuspy/siriuspy/meas/__init__.py
@@ -1,0 +1,3 @@
+from .util import ProcessImage
+
+del util
