+++ siriuspy/siriuspy/meas/liemit/__init__.py
@@ -1,0 +1,3 @@
+from .main import CalcEmmitance, MeasEmmitance
+
+del main
