+++ siriuspy/siriuspy/meas/lienergy/__init__.py
@@ -1,0 +1,5 @@
+from .main import CalcEnergy, MeasEnergy
+
+del main
+
+__all__ = ('main', 'csdev')
